+++ NEWS.md
@@ -1,6 +1,10 @@
 corpus 0.9.3.9000
 =================
 
+### BUG FIXES
+
+  * Fix bug in `print.corpus_frame(,row.names = FALSE)`.
+
 corpus 0.9.3 (2017-10-02)
 =========================